pyproject.toml
[tool.poetry]
name = "resumegenerator"
version = "0.1.0"
description = ""
authors = ["Amulya Paritosh <amulya1171@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pylatex = "^1.4.2"
pydantic = {extras = ["email"], version = "^2.5.3"}
pydantic-extra-types = "^2.4.1"
phonenumbers = "^8.13.28"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0.post1"}
pytest = "^8.0.0"
httpx = "^0.26.0"
faker = "^22.6.0"
firebase-admin = "^6.5.0"
jinja2 = "^3.1.3"
google-cloud-datastore = "^2.19.0"

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
